+++ stellaris-modding-support/src/main/kotlin/com/windea/plugin/idea/stellaris/localization/editor/StellarisLocalizationGoToDeclarationHandler.kt
@@ -1,6 +1,4 @@
-@file:Suppress("UNCHECKED_CAST")
-
-package com.windea.plugin.idea.stellaris.localization.navigation
+package com.windea.plugin.idea.stellaris.localization.editor
 
 import com.intellij.codeInsight.navigation.actions.*
 import com.intellij.openapi.editor.*
@@ -8,22 +6,16 @@ import com.intellij.psi.*
 import com.windea.plugin.idea.stellaris.*
 import com.windea.plugin.idea.stellaris.localization.psi.*
 
-//定位定义：
-//TODO 属性 - 当前文件
-//TODO 属性 - 当前项目
-//TODO 属性 - 游戏目录
-//TODO 属性 - mod目录
-
-class StellarisLocalizationGoToDeclarationHandler : GotoDeclarationHandler{
+class StellarisLocalizationGoToDeclarationHandler : GotoDeclarationHandler {
 	override fun getGotoDeclarationTargets(sourceElement: PsiElement?, offset: Int, editor: Editor?): Array<out PsiElement?>? {
 		return when(sourceElement) {
 			is StellarisLocalizationProperty -> {
-				val name = sourceElement.name?:return null
+				val name = sourceElement.name ?: return null
+				val locale = (sourceElement.containingFile as? StellarisLocalizationFile)?.stellarisLocale
 				//查找当前项目
-				findLocalizationProperties(name, sourceElement.project).toTypedArray()
+				findLocalizationProperties(name, sourceElement.project, locale).toTypedArray()
 			}
 			else -> null
 		}
 	}
 }
-